backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
pytest==7.4.3
google-generativeai==0.5.4
pinecone-client==2.2.4
asyncpraw==7.7.1
python-dotenv==1.0.0
python-multipart==0.0.6